Split `PurchasesTests` step 7: extracted `PurchasesDeferredPurchasesTests`
Follow up to #1609.
I've also cleaned up the initialization of these tests.
